refactor(frontend/models): add specific folder for mock data, adjust import path
